examples/youtube-summarizer/.claude/skills/youtube-summarize/scripts/requirements.txt
yt-dlp>=2024.0.0
openai-whisper>=20231117
